获取otp验证码后端与前端代码实现，用户注册后端代码实现

+++ src/main/java/com/miaoshaproject/service/model/UserModel.java
@@ -31,7 +31,7 @@ public class UserModel {
         return gender;
     }
 
-    public void setGender(Byte gender) {
+    public void setGender(Integer gender) {
         this.gender = gender;
     }
